requirements.txt
# AI Partner Churn Predictor - Python Dependencies
# Day 1: Data Foundation

# Core data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Machine Learning (Day 2-3)
scikit-learn>=1.3.0
lightgbm>=4.0.0
# xgboost>=2.0.0  # Optional, LightGBM preferred

# Causal Inference (Day 3)
causalml>=0.15.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Dashboard (Day 4)
streamlit>=1.28.0
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
# Backend (Day 5)
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
supabase>=2.0.0
